Issue 5-26 120pm commit - Done w integrating diff customer levels into demand draw

+++ aiModule/CafeEatcon/Assets/scripts/AvgCostScript.cs
@@ -5,7 +5,7 @@ using UnityEngine.UI;
 
 public class AvgCostScript : MonoBehaviour {
 
-	public float avgMealCost;
+	public double avgMealCost;
 	public double maxCustomers;
 	public double potentialCustomers;
 	public double potentialIncome;
